created folder for the whole game (...Game/GameComponent/PhaserGame), changed some names, created Game.tsx for injection point of game logic
