+++ front-vite/src/Pages/Game/GameComponent/PhaserGame/Scenes/Error.tsx
